+++ CHANGELOG.md
@@ -31,8 +31,8 @@
 
 1.2.0.0
 -------
-* it is now possible to choose which Jack polynomial to get or evaluate, `J`, `C`, `P` or `Q` 
-(the previous versions returned `J` only)
+* it is now possible to choose which Jack polynomial to get or evaluate, 
+`J`, `C`, `P` or `Q` (the previous versions returned `J` only)
 
 * it is now possible to get Jack polynomials with a symbolic Jack parameter